requirements.txt
yt-dlp
openai
pyannote.audio
torch
python-dotenv
tqdm
requests 